+++ src/components/modals/index.ts
@@ -1,4 +1,7 @@
+export { BaseCategoryOperationModal } from "./base-category-operation-modal";
 export { CalendarSelectModal } from "./calendar-select-modal";
+export { CategoryDeleteModal } from "./category-delete-modal";
+export { CategoryRenameModal } from "./category-rename-modal";
 export { CalendarIntegrationDeleteEventsModal, DeleteRecurringEventsModal } from "./delete-physical-events-modal";
 export { EventCreateModal } from "./event-create-modal";
 export { EventEditModal } from "./event-edit-modal";